pyproject.toml
[tool.poetry]
name = "clova-tes"
version = "0.1.0"
description = ""
authors = ["Fillsogood <jdd05273@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
requests = "^2.32.3"
python-multipart = "^0.0.17"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.4"
pydantic = "^2.10.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
